+++ probe.hpp
@@ -72,7 +72,7 @@ public:
 
     int makeProbe( cfpix &cpix, int nx, int ny, double xp, double yp,
     double wavlen, double k2max, double pixel, double** aber,
-    int ismoth, double kx[], double kx2[], double ky[], double ky2[] );
+    int ismoth, float kx[], float kx2[], float ky[], float ky2[] );
 
     double prbSize( float** pixsq, int nx, int ny,
         double xp, double yp, double ax, double by );
